mast: move module into a separate header
Just a bit of declutter. In the mAST, module is not a declaration so it
makes sense to move it into another header.
