+++ lang/CMakeLists.txt
@@ -24,8 +24,9 @@ aw_add_library(awscript-lang STATIC
 		include/aw/script/ast/statement.h
 		include/aw/script/ast/type.h
 		include/aw/script/ast/middle/declaration.h
-		include/aw/script/ast/middle/statement.h
 		include/aw/script/ast/middle/expression.h
+		include/aw/script/ast/middle/module.h
+		include/aw/script/ast/middle/statement.h
 		include/aw/script/diag/diagnostic.h
 		include/aw/script/diag/diagnostics_engine.h
 		include/aw/script/diag/error_t.h
