+++ lang/include/aw/script/ast/middle/statement.h
@@ -40,7 +40,7 @@ struct while_statement {
 	std::unique_ptr<statement>  loop_body;
 };
 
-class declaration;
+struct declaration;
 struct decl_statement {
 	decl_statement(declaration&& decl);
 	decl_statement(std::unique_ptr<declaration> decl);
