+++ lang/source/semantic/resolve_references.h
@@ -3,9 +3,12 @@
 
 #include "context.h"
 
-#include <aw/script/diag/diagnostics_engine.h>
+#include "aw/script/diag/diagnostics_engine.h"
 
 namespace aw::script {
+namespace middle {
+struct module;
+} // namespace middle;
 
 void resolve_references(context& ctx, diagnostics_engine& diag, middle::module& mod);
